Add files via upload
This shell script allows a user without access to a scheduler to control the frequency of program execution. It uses conditional logic, a sleep timer and a loop to process SAS code every 15 minutes. To setup the program, you need to know what time the scheduled job starts, then wrap this code around the scheduled program.